refactor: User 관련 protected api 요청 모듈화
