+++ server/routers/userRouter.js
@@ -1,9 +1,15 @@
 const express = require("express");
-const loginUser = require("../controllers/userController/logInUser");
+const logInUser = require("../controllers/userController/logInUser");
 const createUser = require("../controllers/userController/createUser");
+const logOutUser = require("../controllers/userController/logOutUser");
+const withdrawUser = require("../controllers/userController/withdrawUser");
 const userRouter = express.Router();
 
 userRouter.post("/create", createUser);
-userRouter.post("/login", loginUser);
+userRouter.post("/login", logInUser);
+
+// protected
+userRouter.post("/signout", logOutUser);
+userRouter.post("/:userId/withdraw", withdrawUser);
 
 module.exports = userRouter;
